feat:Enemy Patrolling AI
Incorporate Enemy Patrolling for 3 State AI.

Works by having an array of points to patrol to. When enemy has a point to get to, onAPoint is set to true as to not overload the function. Once they reach their destination, Boolean is set to false and cycle is repeated.

Next to implement is Alert state; will go to player's last known position if outside the viewRange and stay there for a bit until going back to Patrol state.